Fix for compile warnings

See merge request inference-engine/kmb-plugin!1433

+++ src/mcmCompiler/include/mcm/tensor/shape.hpp
@@ -55,13 +55,13 @@ namespace mv
         std::size_t ndims() const;
         std::size_t totalSize() const;
         bool isFlat() const;
-        std::size_t& operator[](int ndim);
+        std::size_t& operator[](long long int ndim);
 
         static Shape broadcast(const Shape& s1, const Shape& s2);
         static Shape augment(const Shape& s, std::size_t ndims);
         static Shape augment_major(const Shape& s, std::size_t ndims);
 
-        const std::size_t& operator[](int ndim) const;
+        const std::size_t& operator[](long long int ndim) const;
         const std::size_t& operator[](const std::string& ndim) const;
 
         Shape& operator=(const Shape& other);
